UPSTREAM: mb/google/guybrush: PCIe GPIOs - enable enables, disable resets
To train PCIe devices, the devices need to be enabled and taken out of
reset.  This patch does the bare minimum needed to train PCIe.  It is
not intended to handle timings, which will be addressed later.

Copy the enables for WWAN & WLAN into early GPIO Init so that they're
enabled before FSP-M runs and trains the PCIe busses.

Again, this patch is the minimum to let the FSP train the PCIe busses.

BUG=b:182202136
TEST=Boot guybrush from NVME.

Change-Id: I8a290ec2527e9b3398e3a50f1443b573c377f3a2
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 3db49929bd7d755aad920ae6b8757bba57864989
Original-Signed-off-by: Martin Roth <martinroth@chromium.org>
Original-Change-Id: I3807e02de1e9ae40b0a4162217afd6aabb5b04ed
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/52115
Original-Reviewed-by: Raul Rangel <rrangel@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2815005
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>