UPSTREAM: soc/intel/alderlake: rename CONFIG_MAX_PCIE_CLOCKS to CONFIG_MAX_PCIE_CLOCK_SRC
CONFIG_MAX_PCIE_CLOCKS renamed to MAX_PCIE_CLOCK_SRC to make it clear that this config
is for the number of PCIe Clock sources available which is different from PCIe clock reqs.
This is more relevant in alderlake, as the number clock source and clock reqs differ.
However since this is a better name, renaming it throughout the soc/intel tree.

BUG=none
BRANCH=none
TEST=none

Change-Id: I200e64d32f447806c09031119a5abf7f9fb8da2e
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: a979460614db63212f04658a592bd883f8e28b80
Original-Signed-off-by: Rizwan Qureshi <rizwan.qureshi@intel.com>
Original-Change-Id: I747c94331b68c4ec0b6b5a04149856a4bb384829
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/52194
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2843706
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>