UPSTREAM: cbfs: mcache: Fix size calculation for perfectly full cache
cbfs_mcache_real_size() has a subtle flaw: when the cache is perfectly
full to the end (so that the termination token sits exactly at the end
of the available space), the loop counting the size ends prematurely.
This means that when migrating the cache to CBMEM the terminating token
is not copied, which isn't actually noticeable unless you're looking for
a file that's not in the cache (because it doesn't exist or because not
all files fit when building).

This patch fixes the problem and slightly changes the error message for
when a cache isn't terminated (to make it more clear that this is a
different condition from a "normal" cache overflow that can happen when
building if there's not enough room to fit all files).

BUG=none
BRANCH=none
TEST=none

Change-Id: I8494fce0231e3ff42f9efe110e25692e1a78f766
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: b571846ea4e69981dc593faa82c861cdd6ece078
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: I8d89e7dadc958f97b173b3a2352f2010c8a3d1d5
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/52200
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2817772
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>