UPSTREAM: commonlib/region: Turn addrspace_32bit into a more official API
We had the addrspace_32bit rdev in prog_loaders.c for a while to help
represent memory ranges as an rdev, and we've found it useful for a
couple of things that have nothing to do with program loading. This
patch moves the concept straight into commonlib/region.c so it is no
longer anchored in such a weird place, and easier to use in unit tests.
Also expand the concept to the whole address space (there's no real need
to restrict it to 32 bits in 64-bit environments) and introduce an
rdev_chain_mem() helper function to make it a bit easier to use. Replace
some direct uses of struct mem_region_device with this new API where it
seems to make sense.

BUG=none
BRANCH=none
TEST=none

Change-Id: Id3ec174c3a745dcac343910b713e2314497e784a
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: c893197352acc9b53c1beef5082cbc0271f63688
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: Ie4c763b77f77d227768556a9528681d771a08dca
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/52533
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2843670
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>