UPSTREAM: mb/google/mancomb: PCIe GPIOs - enable enables, disable resets
To train PCIe devices, the devices need to be enabled and taken out of
reset.  This patch does the bare minimum needed to train PCIe.  It is
not intended to handle timings, which will be addressed later.

Copy the enables for WLAN into early GPIO Init so that they're
enabled before FSP-M runs and trains the PCIe busses.

Again, this patch is the minimum to let the FSP train the PCIe busses.

BUG=b:182202136
TEST=Boot guybrush from NVME.

Change-Id: I89faebb332a22989667aa95615dffecb05d264a4
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: c20f33960cbf57aa33a1b8ea32799511dabbb18c
Original-Signed-off-by: Ivy Jian <ivy_jian@compal.corp-partner.google.com>
Original-Change-Id: I5e3e9fe21f44b832e26b0942759ae2ec96ec6c82
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/52621
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: EricR Lai <ericr_lai@compal.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2850911
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>