UPSTREAM: device: Switch pci_dev_is_wake_source to take pci_devfn_t
With the recent switch to SMM module loader v2, the size of the SMM for
module google/volteer increased to above 64K in size, and thus failed to
install the permanent SMM handler. Turns out, the devicetree is all
pulled into the SMM build because of elog, which calls
`pci_dev_is_wake_source`, and is the only user of `struct device` in
SMM. Changing this function to take a pci_devfn_t instead allows the
linker to remove almost the entire devicetree from SMM (only usage left
is when disabling HECI via SMM).

BUG=b:186661594
TEST=Verify loaded program size of `smm.elf` for google/volteer is
almost ~50% smaller.

Change-Id: Ica5cc21801669359e673ee0a7654a16a077dc461
Original-Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Commit-Id: 93982c3a6ede0bdb8dd0e94597e4de7b09406947
Original-Change-Id: I4c39e5188321c8711d6479b15065e5aaedad8f38
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/52765
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2867791