UPSTREAM: drivers/i2c/generic: Set S0W to D3hot for wake device
If device is supported as a wake source, _S0W should be set to D3hot.
This ensures that the device is put into D3hot by the OSPM.

Power resource(PRIC) for the device is listed in both _PR0 and _PR3. Thus, it ensures that the OSPM does not turn off power resource when device is put into D0 and D3hot. Hence, it is capable of waking the system from D3hot state. However, if it is put into D3cold, then the power resource is turned off by the OSPM.

The devices we are currently looking at for touchscreen/touchpad
do not really support auxiliary power and so do not support wake from D3cold.

BUG=b:186070097
TEST=build and check device wake state _S0W set to 3 in ssdt table.

Change-Id: Ia2ca91f50ac888850af97ad144e8c7fa2f90133e
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 1282b007be8c4ee2c6193a71f66b66ef989be4b3
Original-Change-Id: I34e4b2350875530d3337be700276bcc4fb1f810a
Original-Signed-off-by: Tony Huang <tony-huang@quanta.corp-partner.google.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/52847
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2874898
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>