UPSTREAM: soc/intel/{adl,tgl,jsl}: Add smihandler_soc_disable_busmaster
If a power button SMI is triggered between where it is currently
enabled and before FSP-S exits, when the SMI handler disables
bus mastering for all devices, it inadvertently also disables
the PMC's I/O decoding, so the register write to actually go into
S5 does not succeed, and the system hangs.

This can be solved by skipping the PMC when disabling bus
mastering in the SMI handler, for which a callback,
smihandler_soc_disable_busmaster is provided.

BUG=b:186194102, b:186815114
TEST=Power on the system and pressing power button repeatedly doesn't
     cause the system hang during shutdown.

Change-Id: If05f53ea1fffda0b7b9f3660e6c7fb8d92f84a03
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 7b7b33e3a615b0581df9e8e371fb218e97ba2a51
Original-Change-Id: I1cf5cf91ebad4a49df6679e01fc88ff60c81526c
Original-Signed-off-by: Kane Chen <kane.chen@intel.corp-partner.google.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/52873
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2878875
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>