UPSTREAM: mb/google/guybrush: Switch eSPI ALERT# to in-band
Using the push-pull alert was causing leakages when in S0i3. This is
because the EC drives ALERT#, so when the AP enters S0i3, the extra
current leaks into the SoC and ends up turning on the power regulators.
By using in-band ALERT#, the EC no longer drives this pin high, thus
fixing the leak. We could also have used an open drain alert, but the
rise time is less than ideal.

BUG=b:187122344, b:186135022
TEST=Measure S0i3 power on guybrush and validate it's no longer high.

Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Commit-Id: f33f857fbe8e190c2972d4c466a3b7f96e03cd82
Original-Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Original-Change-Id: I6de771aeda8feca062652f0ea9eb57d31cb68562
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/52955
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Karthik Ramasubramanian <kramasub@google.com>
Original-Reviewed-by: Felix Held <felix-coreboot@felixheld.de>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-by: Rob Barnes <robbarnes@google.com>
Change-Id: I6f3b3786267969a63390a3c1482ffbafff99de98
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2879367
Reviewed-by: Martin Roth <martinroth@google.com>
Tested-by: Martin Roth <martinroth@google.com>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>