UPSTREAM: drivers/sn65dsi86: Switch EDID reading to use "indirect mode"
The SN65DSI86 eDP bridge supports two ways to read the EDID: for now
we've been using "direct mode", which works by basically making the
bridge I2C device listen to another chip address besides its own and
proxy all requests received there directly to the eDP AUX channel. The
great part about that mode is that it is super easy and hassle-free to
use. The not so great part about it is that it doesn't work: for EDID
extensions, the last byte (which happens to contain the checksum) is
somehow always read as zero. We presume this is a hardware bug in the
bridge part.

The other, much more annoying way is "indirect mode", where each byte
transmitted over the AUX channel has to be manually set up in the I2C
registers of the bridge, just like we're already doing with DPCD
transactions. Thankfully, we can reuse most of the DPCD code for this so
it's not a lot of extra code. It's a bit slower but not as much as you'd
expect (26ms instead of 18ms on my board), and the difference is not
very relevant compared to common total times for display init.

Also, some of the (previously unused) enum definitions for the AUX_CMD
mode field of the bridge had just been plain wrong for some reason, and
needed to be fixed to make this work.

Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: I65f80193380d3c3841f9f5c26897ed672f45e15a
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/52959
Original-Reviewed-by: Douglas Anderson <dianders@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: I16025d150e9bcddd2adbd8a4670454c8a7715350
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2881142
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>