UPSTREAM: mb/google/volteer: adjust the size for RO/RW mcache
The mcache is overflowed in the latest build. In order
to fix the mcache overflow, we increase the mcache size
to 0x4000 and adjust the percentage to 50% for the ro/rw
mcache. This change is for all of the volteer variants
as we see many of the volteer variants which use the
latest bios having the mcache overflow issue.

BUG=b:187095474, b:187095765, b:187234881, b:162052593
TEST=no mcache overflow in the bios log

Change-Id: Ifd723634e490bf0fd0fb8069c7df9cbc5ee524f8
Original-Change-Id: If9552bc9fa5d36b1ca662c9da030ae7b137b60a8
Original-Commit-Id: 509b0a97f6a44ea85a8a37087a03526dd513a88f
Original-Signed-off-by: Zhuohao Lee <zhuohao@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/53908
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2881374