UPSTREAM: soc/amd/common/block/pci: Implement acpigen_write_pci_{GNB,FCH}_PRT
This is loosely based off of picasso/pcie_gpp.c. This version uses the
acpigen_write_PRT_X methods to write the actual records. There are also
two functions, 1 for using the GNB, and one for using the FCH. The FCH
one is useful when the GNB IO-APIC has not been initialized.

BUG=b:184766519
TEST=Dump guybrush ACPI and verify it looks correct

Change-Id: Ie5447d9795ac53fcf13464ef292a86a9b239d360
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 7502e10fdfe4085c5f7d03b0b29d0f7fb9b390fe
Original-Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Original-Change-Id: I926430074acb969ceb11fdb60ab56dcf91ac4c76
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/52917
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Felix Held <felix-coreboot@felixheld.de>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2883450
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>