UPSTREAM: vboot/secdata_mock: Make v0 kernel secdata context
The new kernel secdata v1 stores the last read EC hash, and reboots the
device during EC software sync when that hash didn't match the currently
active hash on the EC (this is used with TPM_CR50 to support EC-EFS2 and
pretty much a no-op for other devices). Generally, of course the whole
point of secdata is always that it persists across reboots, but with
MOCK_SECDATA we can't do that. Previously we always happened to somewhat
get away with presenting freshly-reinitialized data for MOCK_SECDATA on
every boot, but with the EC hash feature in secdata v1, that would cause
a reboot loop. The simplest solution is to just pretend we're a secdata
v0 device when using MOCK_SECDATA.

This was encountered on using a firmware built with MOCK_SECDATA but had
EC software sync enabled.

BUG=b:187843114
BRANCH=None
TEST=`USE=mocktpm cros build-ap -b keeby`; Flash keeby device, verify
that DUT does not continuously reboot with EC software sync enabled.

Change-Id: I73e0314a5c5b2938e5f987fb6820e2029eb13c35
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 3c79777cd62a66ede560b1069565fc42a8e0f008
Original-Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Original-Change-Id: Id8e81afcddadf27d9eec274f7f85ff1520315aaa
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/54304
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2909657
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>