UPSTREAM: vboot/secdata_tpm: Rename set_space()
The name `set_space()` seems to imply that it's writing to a TPM space
when actually, the function can create a space and write to it.  This
commit attempts to make that a bit more clear.  Additionally, in order
to use the correct sizes when creating the space, this commit also
refactors the functions slightly to incorporate the vboot context object
such that the correct sizes are used.  The various vboot APIs will
return the size of the created object that we can then create the space
with.

BUG=b:184677625
BRANCH=None
TEST=`emerge-keeby coreboot`

Change-Id: I667073ee717e87487f16a1a9d2cc1fb228962281
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: d87ed2d551c0d0f7ba6f0e874c9496cee09d19a7
Original-Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Original-Change-Id: I80a8342c51d7bfaa0cb2eb3fd37240425d5901be
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/54308
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2900264
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>