UPSTREAM: soc/amd/cezanne/fch: add PCIe GPP clock generator configuration settings
I'm not 100% sure if this should rather be duplicated from Picasso or
commonized. Checked with the docs and this won't be compatible with
Stoneyridge and one future product's PPR lacked the corresponding
register. Some other chip has a compatible register layout, but a
different number of PCIe GPP clock outputs, so the common code would
need to use some SoC-dependent defines and possibly a SoC-specific
lookup table for the mapping which is also not that great.

TEST=Checked Cezanne PPR

Change-Id: Iadcca05e34eef5e0b3bc9f58c4c0365d580d2a77
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 224b578420a5d42e16ec6a8285971d34d8cdafac
Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Change-Id: I6b6d0cb8d7eb0288d8a18fcb975dc377b2c6846a
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/54685
Original-Reviewed-by: Raul Rangel <rrangel@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2909671
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>